Update RCTContacts.m to use correct UIBarButtonItemStyle constant
`openExistingContact:...` was creating a `UIBarButtonItem` with a `style` of `UIBarButtonSystemItemCancel`.  However `UIBarButtonSystemItemCancel` is not a style, it is a type of item.  Instead we should use `UIBarButtonItemStyleDone`

This will avoid raising a `-Wenum-conversion` error and fixes the bug for the style ending up being the deprecated `UIBarButtonItemStyleBordered` (`0x1`) by accident.